[refactor] Confilict 해결
